+++ src/css/theme.css
@@ -177,7 +177,7 @@ input#img16 {
 }
 
 /*
-***********PostDetailsPage PageLayout****************
+***********PostDetail PageLayout****************
 */
 .post-container {
     padding: 0 15px;
