+++ src/index.js
@@ -14,16 +14,16 @@ import {auth} from 'firebase/index';
 
 //PageLayout Components
 import Home from './components/pages/Home/Home.jsx';
-import NotFound from './components/pages/404/notfound.jsx';
-import PostDetailsPage from './components/pages/PostDetailsPage/PostDetailsPage';
+import Notfound from './components/pages/404/Notfound.jsx';
+import PostDetail from './components/pages/PostDetail/PostDetail';
 import NewPost from './components/pages/CreatePost/CreatePost';
 import Login from './components/pages/Login/Login';
-import Register from './components/pages/register/Register.jsx';
-import About from './components/pages/about/about.jsx';
+import Register from './components/pages/Register/Register.jsx';
+import About from './components/pages/About/about.jsx';
 
 //Template parts
-import Header from './components/template/header/Header';
-import Footer from './components/template/footer/Footer';
+import Header from './components/template/Header/Header';
+import Footer from './components/template/Footer/Footer';
 
 //Styles
 import './css/theme.css'
@@ -37,7 +37,7 @@ const Root = (props) => {
                 <Header {...{user}}/>
                 <div className="content">
                     <Switch>
-                        <Route path="/post/:postId" render={(params) => <PostDetailsPage {...{params, user}} />} />
+                        <Route path="/post/:postId" render={(params) => <PostDetail {...{params, user}} />} />
                         <Route exact path="/category/:category" render={(params) => <Home {...{params, user}}/>} />
                         <Route exact path="/category/" render={() => <Redirect to="/"/>}/>
                         <Route exact path="/login" render={() => <Login {...{user}}/>} />
@@ -45,7 +45,7 @@ const Root = (props) => {
                         <Route exact path="/register" render={() => <Register {...{user}}/>}/>
                         <Route exact path="/about" component={About}/>
                         <Route exact path="/" render={() => <Home {...{user}}/>}/>
-                        <Route component={NotFound}/>
+                        <Route component={Notfound}/>
                     </Switch>
                 </div>
                 <Footer/>
