+++ src/registerServiceWorker.js
@@ -1,11 +1,11 @@
-// In production, we register a service worker to serve assets from local cache.
+// In production, we Register a service worker to serve assets from local cache.
 
 // This lets the app load faster on subsequent visits in production, and gives
 // it offline capabilities. However, it also means that developers (and users)
 // will only see deployed updates on the "N+1" visit to a page, since previously
 // cached resources are updated in the background.
 
-// To learn more about the benefits of this model, read https://goo.gl/KwvDNy.
+// To learn more About the benefits of this model, read https://goo.gl/KwvDNy.
 // This link also includes instructions on opting out of this behavior.
 
 const isLocalhost = Boolean(
@@ -45,7 +45,7 @@ export default function register() {
           );
         });
       } else {
-        // Is not local host. Just register service worker
+        // Is not local host. Just Register service worker
         registerValidSW(swUrl);
       }
     });
